injectable variables and functions
