+++ interpreter/context/option.go
@@ -43,9 +43,3 @@ func WithRequest(r *config.RequestConfig) Option {
 		c.OverrideRequest = r
 	}
 }
-
-func WithTesting(isTesting bool) Option {
-	return func(c *Context) {
-		c.IsTesting = isTesting
-	}
-}
