+++ interpreter/variable/all.go
@@ -29,14 +29,12 @@ import (
 // see: https://docs.google.com/spreadsheets/d/1uV9yRdOpMQxyMm50VRBA1MDQIkcgg_cwPAgslhqatgw/edit#gid=258583317
 type AllScopeVariables struct {
 	Variable
-	ctx     *context.Context
-	testing *TestingVariables
+	ctx *context.Context
 }
 
 func NewAllScopeVariables(ctx *context.Context) *AllScopeVariables {
 	return &AllScopeVariables{
-		ctx:     ctx,
-		testing: NewTestingVariables(ctx),
+		ctx: ctx,
 	}
 }
 
@@ -568,8 +566,10 @@ func (v *AllScopeVariables) Get(s context.Scope, name string) (value.Value, erro
 		return val, nil
 	}
 
-	if tv := v.testing.GetTestingVariable(name); tv != nil {
-		return tv, nil
+	if injectedVariable != nil {
+		if val, err := injectedVariable.Get(v.ctx, s, name); err == nil {
+			return val, nil
+		}
 	}
 
 	return value.Null, errors.WithStack(fmt.Errorf(
@@ -713,6 +713,12 @@ func (v *AllScopeVariables) Set(s context.Scope, name, operator string, val valu
 	if match := requestHttpHeaderRegex.FindStringSubmatch(name); match != nil {
 		v.ctx.Request.Header.Set(match[1], val.String())
 		return nil
+	}
+
+	if injectedVariable != nil {
+		if err := injectedVariable.Set(v.ctx, s, name, operator, val); err == nil {
+			return nil
+		}
 	}
 
 	return errors.WithStack(fmt.Errorf(
