+++ tester/function/assert_ends_with.go
@@ -1,4 +1,4 @@
-package testings
+package function
 
 import (
 	"strings"
