+++ tester/function/assert_equal.go
@@ -1,4 +1,4 @@
-package testings
+package function
 
 import (
 	"fmt"
