+++ tester/function/assert_match.go
@@ -1,4 +1,4 @@
-package testings
+package function
 
 import (
 	"regexp"
@@ -8,31 +8,31 @@ import (
 	"github.com/ysugimoto/falco/interpreter/value"
 )
 
-const Assert_not_match_lookup_Name = "assert"
+const Assert_match_lookup_Name = "assert"
 
-var Assert_not_match_lookup_ArgumentTypes = []value.Type{value.StringType, value.StringType}
+var Assert_match_lookup_ArgumentTypes = []value.Type{value.StringType, value.StringType}
 
-func Assert_not_match_lookup_Validate(args []value.Value) error {
+func Assert_match_lookup_Validate(args []value.Value) error {
 	if len(args) < 2 || len(args) > 3 {
-		return errors.ArgumentNotInRange(Assert_not_match_lookup_Name, 2, 3, args)
+		return errors.ArgumentNotInRange(Assert_match_lookup_Name, 2, 3, args)
 	}
 
-	for i := range Assert_not_match_lookup_ArgumentTypes {
-		if args[i].Type() != Assert_not_match_lookup_ArgumentTypes[i] {
-			return errors.TypeMismatch(Assert_not_match_lookup_Name, i+1, Assert_not_match_lookup_ArgumentTypes[i], args[i].Type())
+	for i := range Assert_match_lookup_ArgumentTypes {
+		if args[i].Type() != Assert_match_lookup_ArgumentTypes[i] {
+			return errors.TypeMismatch(Assert_match_lookup_Name, i+1, Assert_match_lookup_ArgumentTypes[i], args[i].Type())
 		}
 	}
 
 	if len(args) == 3 {
 		if args[2].Type() != value.StringType {
-			return errors.TypeMismatch(Assert_not_match_lookup_Name, 3, value.StringType, args[2].Type())
+			return errors.TypeMismatch(Assert_match_lookup_Name, 3, value.StringType, args[2].Type())
 		}
 	}
 	return nil
 }
 
-func Assert_not_match(ctx *context.Context, args ...value.Value) (value.Value, error) {
-	if err := Assert_not_match_lookup_Validate(args); err != nil {
+func Assert_match(ctx *context.Context, args ...value.Value) (value.Value, error) {
+	if err := Assert_match_lookup_Validate(args); err != nil {
 		return nil, errors.NewTestingError(err.Error())
 	}
 
@@ -50,12 +50,12 @@ func Assert_not_match(ctx *context.Context, args ...value.Value) (value.Value, e
 		return nil, errors.NewTestingError("Invalid regexp string provided: %s", actual.Value)
 	}
 	ret := &value.Boolean{Value: re.MatchString(expect.Value)}
-	if ret.Value {
+	if !ret.Value {
 		if message != "" {
 			return ret, errors.NewAssertionError(message)
 		}
 		return ret, errors.NewAssertionError(
-			"String match: value %s should not match against %s",
+			"String unmatched: value %s does not match against %s",
 			expect.Value,
 			actual.Value,
 		)
