+++ tester/function/assert_not_contains_test.go
@@ -1,4 +1,4 @@
-package testings
+package function
 
 import (
 	"testing"
@@ -10,7 +10,7 @@ import (
 	"github.com/ysugimoto/falco/interpreter/value"
 )
 
-func Test_Assert_ends_with(t *testing.T) {
+func Test_Assert_not_contains(t *testing.T) {
 
 	tests := []struct {
 		args   []value.Value
@@ -20,14 +20,14 @@ func Test_Assert_ends_with(t *testing.T) {
 		{
 			args: []value.Value{
 				&value.String{Value: "foobarbaz"},
-				&value.String{Value: "baz"},
+				&value.String{Value: "bat"},
 			},
 			expect: &value.Boolean{Value: true},
 		},
 		{
 			args: []value.Value{
 				&value.String{Value: "foobarbaz"},
-				&value.String{Value: "bat"},
+				&value.String{Value: "bar"},
 			},
 			expect: &value.Boolean{Value: false},
 			err:    &errors.AssertionError{},
@@ -43,7 +43,7 @@ func Test_Assert_ends_with(t *testing.T) {
 		{
 			args: []value.Value{
 				&value.String{Value: "foobarbaz"},
-				&value.String{Value: "bat"},
+				&value.String{Value: "bar"},
 				&value.String{Value: "custom_message"},
 			},
 			expect: nil,
@@ -54,7 +54,7 @@ func Test_Assert_ends_with(t *testing.T) {
 	}
 
 	for i := range tests {
-		_, err := Assert_ends_with(
+		_, err := Assert_not_contains(
 			&context.Context{},
 			tests[i].args...,
 		)
@@ -64,7 +64,7 @@ func Test_Assert_ends_with(t *testing.T) {
 			cmpopts.IgnoreFields(errors.AssertionError{}, "Message"),
 			cmpopts.IgnoreFields(errors.TestingError{}, "Message"),
 		); diff != "" {
-			t.Errorf("Assert_ends_with()[%d] error: diff=%s", i, diff)
+			t.Errorf("Assert_not_contains()[%d] error: diff=%s", i, diff)
 		}
 	}
 }
