+++ tester/function/assert_not_match_test.go
@@ -1,4 +1,4 @@
-package testings
+package function
 
 import (
 	"testing"
@@ -10,7 +10,7 @@ import (
 	"github.com/ysugimoto/falco/interpreter/value"
 )
 
-func Test_Assert_contains(t *testing.T) {
+func Test_Assert_not_match(t *testing.T) {
 
 	tests := []struct {
 		args   []value.Value
@@ -20,14 +20,14 @@ func Test_Assert_contains(t *testing.T) {
 		{
 			args: []value.Value{
 				&value.String{Value: "foobarbaz"},
-				&value.String{Value: "bar"},
+				&value.String{Value: ".*bat.*"},
 			},
 			expect: &value.Boolean{Value: true},
 		},
 		{
 			args: []value.Value{
 				&value.String{Value: "foobarbaz"},
-				&value.String{Value: "bat"},
+				&value.String{Value: ".*bar.*"},
 			},
 			expect: &value.Boolean{Value: false},
 			err:    &errors.AssertionError{},
@@ -43,7 +43,7 @@ func Test_Assert_contains(t *testing.T) {
 		{
 			args: []value.Value{
 				&value.String{Value: "foobarbaz"},
-				&value.String{Value: "bat"},
+				&value.String{Value: ".*bar.*"},
 				&value.String{Value: "custom_message"},
 			},
 			expect: nil,
@@ -51,10 +51,18 @@ func Test_Assert_contains(t *testing.T) {
 				Message: "custom_message",
 			},
 		},
+		{
+			args: []value.Value{
+				&value.String{Value: "foobarbaz"},
+				&value.String{Value: "^++a"},
+			},
+			expect: nil,
+			err:    &errors.TestingError{},
+		},
 	}
 
 	for i := range tests {
-		_, err := Assert_contains(
+		_, err := Assert_not_match(
 			&context.Context{},
 			tests[i].args...,
 		)
@@ -64,7 +72,7 @@ func Test_Assert_contains(t *testing.T) {
 			cmpopts.IgnoreFields(errors.AssertionError{}, "Message"),
 			cmpopts.IgnoreFields(errors.TestingError{}, "Message"),
 		); diff != "" {
-			t.Errorf("Assert_contains()[%d] error: diff=%s", i, diff)
+			t.Errorf("Assert_not_match()[%d] error: diff=%s", i, diff)
 		}
 	}
 }
