+++ tester/function/assert_starts_with.go
@@ -1,4 +1,4 @@
-package testings
+package function
 
 import (
 	"strings"
