+++ tester/function/assert_strict_equal.go
@@ -1,4 +1,4 @@
-package testings
+package function
 
 import (
 	"github.com/ysugimoto/falco/interpreter/context"
@@ -6,22 +6,22 @@ import (
 	"github.com/ysugimoto/falco/interpreter/value"
 )
 
-const Assert_not_strict_equal_lookup_Name = "assert"
+const Assert_strict_equal_lookup_Name = "assert"
 
-func Assert_not_strict_equal_lookup_Validate(args []value.Value) error {
+func Assert_strict_equal_lookup_Validate(args []value.Value) error {
 	if len(args) < 2 || len(args) > 3 {
-		return errors.ArgumentNotInRange(Assert_not_strict_equal_lookup_Name, 2, 3, args)
+		return errors.ArgumentNotInRange(Assert_strict_equal_lookup_Name, 2, 3, args)
 	}
 	if len(args) == 3 {
 		if args[2].Type() != value.StringType {
-			return errors.TypeMismatch(Assert_not_strict_equal_lookup_Name, 3, value.StringType, args[2].Type())
+			return errors.TypeMismatch(Assert_strict_equal_lookup_Name, 3, value.StringType, args[2].Type())
 		}
 	}
 	return nil
 }
 
-func Assert_not_strict_equal(ctx *context.Context, args ...value.Value) (value.Value, error) {
-	if err := Assert_not_strict_equal_lookup_Validate(args); err != nil {
+func Assert_strict_equal(ctx *context.Context, args ...value.Value) (value.Value, error) {
+	if err := Assert_strict_equal_lookup_Validate(args); err != nil {
 		return nil, errors.NewTestingError(err.Error())
 	}
 
@@ -43,5 +43,5 @@ func Assert_not_strict_equal(ctx *context.Context, args ...value.Value) (value.V
 		return &value.Boolean{}, errors.NewAssertionError(message)
 	}
 
-	return assert_not(expect.String(), actual.String(), message)
+	return assert(expect.String(), actual.String(), message)
 }
